+++ R/xyz2R.R
@@ -10,7 +10,7 @@
 #'          In order to make \code{T} align with \code{B}, we first rotate \code{T}
 #'          an angle \code{x} about its x-axis (common axis for both \code{A} and \code{T}).
 #'    \item Secondly, \code{T} is rotated an angle \code{y} about the NEW y-axis of \code{T}.
-#'    \item Finally, code{T} is rotated an angle \code{z} about its NEWEST z-axis.
+#'    \item Finally, \code{T} is rotated an angle \code{z} about its NEWEST z-axis.
 #'          The final orientation of \code{T} now coincides with the orientation of \code{B}.
 #' }
 #' The signs of the angles are given by the directions of the axes and the right hand rule.
@@ -41,10 +41,20 @@ xyz2R <- function(x, y, z) {
   sx <- sin(x)
 
   R_AB <- matrix(
-    c(         cy*cz,          -cy*sz,     sy,
-      sy*sx*cz+cx*sz, -sy*sx*sz+cx*cz, -cy*sx,
-     -sy*cx*cz+sx*sz,  sy*cx*sz+sx*cz, cy*cx
+    c(
+      cy * cz,
+      -cy * sz,
+      sy,
+      sy * sx * cz + cx * sz,
+      -sy * sx * sz + cx * cz,
+      -cy * sx,
+      -sy * cx * cz + sx * sz,
+      sy * cx * sz + sx * cz,
+      cy * cx
     ),
-    nrow = 3, ncol = 3, byrow = TRUE)
+    nrow = 3,
+    ncol = 3,
+    byrow = TRUE
+  )
   R_AB
 }